deployer: Wait 100ms before overwriting existing bootloader
If the power supply isn't fully stable or the reset line is glitchy
then there may be a few reboots before the mcu becomes stable.  If the
deployer runs during this unstable period, it may erase the existing
bootloader, but not be able to fully write the new bootloader.

Wait a 100ms before erasing the existing bootloader in an effort to
avoid this situation.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>